Bring back decoder
