+++ src/network/receiver.hpp
@@ -1,7 +1,7 @@
 #pragma once
 
 #include "channel.hpp"
-#include "encoder/ffmpeg/unit.hpp"
+#include "codec/ffmpeg/unit.hpp"
 
 
 namespace shar {
@@ -10,7 +10,7 @@ class IPacketReceiver {
 public:
   virtual ~IPacketReceiver() {}
 
-  virtual void run(Sender<encoder::ffmpeg::Unit> units) = 0;
+  virtual void run(Sender<codec::ffmpeg::Unit> units) = 0;
   virtual void shutdown() = 0;
 };
