+++ src/network/rtp/sender.hpp
@@ -9,7 +9,7 @@
 
 #include "context.hpp"
 #include "network/sender.hpp"
-#include "encoder/ffmpeg/unit.hpp"
+#include "codec/ffmpeg/unit.hpp"
 #include "channel.hpp"
 #include "packetizer.hpp"
 #include "cancellation.hpp"
@@ -17,7 +17,7 @@
 
 namespace shar::rtp {
 
-using encoder::ffmpeg::Unit;
+using codec::ffmpeg::Unit;
 
 class PacketSender
   : public IPacketSender
