+++ src/network/sender.hpp
@@ -1,7 +1,7 @@
 #pragma once
 
 #include "channel.hpp"
-#include "encoder/ffmpeg/unit.hpp"
+#include "codec/ffmpeg/unit.hpp"
 
 
 namespace shar {
@@ -10,7 +10,7 @@ class IPacketSender {
 public:
   virtual ~IPacketSender() {}
 
-  virtual void run(Receiver<encoder::ffmpeg::Unit> units) = 0;
+  virtual void run(Receiver<codec::ffmpeg::Unit> units) = 0;
   virtual void shutdown() = 0;
 };
